requirements.txt
streamlit
google-generativeai
pandas
langchain-text-splitters
scikit-learn
numpy
PyMuPDF
python-dotenv